# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Modelo de API para transações da Saque e Pague. O parceiro interessado\
    \ deve implementar as operações desejadas para que a SeP consiga se comunicar\
    \ com o mesmo."
  version: "1.19.0"
  title: "Saque e Pague Transacional"
  contact:
    name: "Time Saque e Pague"
    url: "http://www.saqueepague.com.br"
host: "api-saqueepague.sensedia.com"
basePath: "/saqueepague/SepTransaction/1.19.0"
...
schemes:
- "https"
paths:
...
  /consultaConta:
    post:
      tags:
      - "Depósito sem Cartão"
      - "Depósito via Carteira Digital"
      - "Transferência com Cartão"
      - "Transferência sem Cartão"
      description: "Operação de consulta de conta de favorecido, usada previamente\
        \ ao depósito para verificar se a conta é válida e está habilitada para depó\
        sito."
      operationId: "consultaContaPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de consulta de conta."
        required: true
        schema:
          $ref: "#/definitions/consultaContaReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/consultaContaResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "DepsitoSemCarto"
  /consultaLimitesEmprestimo:
    post:
      tags:
      - "Empréstimo com Cartão"
      description: "Operação de consulta dos limites disponíveis para empréstimo.\
        \ Esta operação é válida somente com os dados do cartão."
      operationId: "consultaLimitesEmprestimoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de consulta limites de empréstimo, usando dados do\
          \ cartão."
        required: true
        schema:
          $ref: "#/definitions/consultaLimitesEmprestimoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/consultaLimitesEmprestimoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "EmprstimoComCarto"
...
  /consultaParcelasEmprestimo:
    post:
      tags:
      - "Empréstimo com Cartão"
      description: "Operação de consulta de condições disponíveis de parcelamento\
        \ do empréstimo."
      operationId: "consultaParcelasEmprestimoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de consulta parcelas de empréstimo."
        required: true
        schema:
          $ref: "#/definitions/consultaParcelasEmprestimoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/consultaParcelasEmprestimoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "EmprstimoComCarto"
...
  /consultaTaxas:
    post:
      tags:
      - "Saque via Carteira Digital"
      - "Depósito via Carteira Digital"
      description: "Operação de consulta de taxas dinâmicas, requisitado antes de\
        \ depósito ou saque de carteira de digital."
      operationId: "consultaTaxasPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de consulta de taxas."
        required: true
        schema:
          $ref: "#/definitions/consultaTaxasReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/consultaTaxasResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "SaqueViaCarteiraDigital"
...
  /deposito:
    post:
      tags:
      - "Depósito sem Cartão"
      - "Depósito com Cartão"
      - "Depósito via Carteira Digital"
      description: "Operação de depósito em dinheiro, cartão é opcional nesta operaç\
        ão."
      operationId: "depositoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Objeto da requisição de depósito, cartão é opcional."
        required: true
        schema:
          $ref: "#/definitions/depositoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/depositoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "DepsitoSemCarto"
  /depositoConf:
    post:
      tags:
      - "Depósito sem Cartão"
      - "Depósito com Cartão"
      - "Depósito via Carteira Digital"
      description: "Confirmação de operação de depósito."
      operationId: "depositoConfPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de confirmação de operação de depósito."
        required: true
        schema:
          $ref: "#/definitions/transacConf"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "DepsitoSemCarto"
  /emprestimo:
    post:
      tags:
      - "Empréstimo com Cartão"
      description: "Operação de empréstimo."
      operationId: "emprestimoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de empréstimo."
        required: true
        schema:
          $ref: "#/definitions/emprestimoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/emprestimoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "EmprstimoComCarto"
  /emprestimoConf:
    post:
      tags:
      - "Empréstimo com Cartão"
      description: "Confirmação de operação de empréstimo."
      operationId: "emprestimoConfPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de confirmação de operação de empréstimo."
        required: true
        schema:
          $ref: "#/definitions/transacConf"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "EmprstimoComCarto"
...
  /identificacao:
    post:
      tags:
      - "Saque com Cartão"
      - "Saldo com Cartão"
      - "Extrato com Cartão"
      - "Depósito com Cartão"
      - "Empréstimo com Cartão"
      - "Pagamento com Cartão"
      - "Transferência com Cartão"
      description: "Operação de identificação do cliente e informações adicionais\
        \ de sua conta."
      operationId: "identificacaoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de identificação."
        required: true
        schema:
          $ref: "#/definitions/identificacaoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/identificacaoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "SaqueComCarto"
...
definitions:
...
  cartao:
    type: "object"
    properties:
      modoEntrada:
        type: "string"
        description: "Modo de entrada de leitura do cartão. (000 = sem cartão, 050\
          \ = chip, 800 = chip, se falhar realiza leitura da trilha, 900 = trilha\
          \ completa)."
        enum:
        - "000"
        - "050"
        - "800"
        - "900"
      pan:
        type: "string"
        example: "zhXTyHl5nf2ckysYXXjAgZS5fzS7ybC5xePnWeLEP//bw=="
        description: "PAN/Número do cartão utilizado (dado criptografado)."
        pattern: "^.{10,}$"
      trilha2:
        type: "string"
        example: "mMIrkIrf2pSHccWuaROxoIS+NGcTOWcTWXUtjniTa1BRzPziiE8pWnknAxW9HIgoKdc/Fli2KZwNA=="
        description: "Trilha 2 do cartão utilizado (dado criptografado)."
        pattern: "^.{10,}$"
...
  consultaContaReq:
    type: "object"
    required:
    - "InfConsultaConta"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cartao:
        $ref: "#/definitions/cartao"
      Cripto:
        $ref: "#/definitions/cripto"
      InfConsultaConta:
        $ref: "#/definitions/infConsultaContaReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  consultaContaResp:
    type: "object"
    required:
    - "InfConsultaConta"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfConsultaConta:
        $ref: "#/definitions/infConsultaContaResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      InfConsultaConta:
        solicDoc: "00"
        modalidadeDeposito: "00"
        nomeCliente: "Carl Edward Sagan"
        depositoIdentificado: "00"
        habilitaDeposito: "01"
      Cripto:
        hash: "hash"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
  consultaLimitesEmprestimoReq:
    type: "object"
    required:
    - "Cartao"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cartao:
        $ref: "#/definitions/cartao"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  consultaLimitesEmprestimoResp:
    type: "object"
    required:
    - "InfConsultaLimitesEmprestimo"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfConsultaLimitesEmprestimo:
        $ref: "#/definitions/infConsultaLimitesEmprestimoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      InfConsultaLimitesEmprestimo:
        limiteMinimoEmprestimo: "000000500000"
        limiteMaximoEmprestimo: "000005000000"
        nomeAgenciaBancaria: "Agência bancária"
        nomeTitularConta: "Carl Edward Sagan"
      Cripto:
        hash: "hash"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
  consultaParcelasEmprestimoReq:
    type: "object"
    required:
    - "Cartao"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cartao:
        $ref: "#/definitions/cartao"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  consultaParcelasEmprestimoResp:
    type: "object"
    required:
    - "InfConsultaParcelasEmprestimo"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfConsultaParcelasEmprestimo:
        $ref: "#/definitions/infConsultaParcelasEmprestimoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfConsultaParcelasEmprestimo:
        opcoesParcelas:
        - valorParcela: "000000005000"
          qntParcelas: "5"
        - valorParcela: "000000005000"
          qntParcelas: "5"
        dataPrimeiraParcela: "20181122"
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      Cripto:
        hash: "hash"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
...
  consultaTaxasReq:
    type: "object"
    required:
    - "InfTransacao"
    - "Terminal"
    properties:
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  consultaTaxasResp:
    type: "object"
    required:
    - "InfConsultaTaxas"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfConsultaTaxas:
        $ref: "#/definitions/infConsultaTaxasResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfConsultaTaxas:
        transactions_fees:
        - fees:
          - min_amount: "000"
            max_amount: "10000"
            id: "0201"
            value: "000"
          - min_amount: "000"
            max_amount: "10000"
            id: "0201"
            value: "000"
          name: "deposit"
          id: "01"
        - fees:
          - min_amount: "000"
            max_amount: "10000"
            id: "0201"
            value: "000"
          - min_amount: "000"
            max_amount: "10000"
            id: "0201"
            value: "000"
          name: "deposit"
          id: "01"
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      Cripto:
        hash: "hash"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
...
  depositoReq:
    type: "object"
    required:
    - "InfDeposito"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cartao:
        $ref: "#/definitions/cartao"
      Cripto:
        $ref: "#/definitions/cripto"
      InfDeposito:
        $ref: "#/definitions/infDepositoReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  depositoResp:
    type: "object"
    required:
    - "InfDeposito"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfDeposito:
        $ref: "#/definitions/infDepositoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      Cripto:
        hash: "hash"
      InfDeposito:
        recibo: "@            COMPROVANTE DE DEPOSITO             @------------------------------------------------@\
          \ Agencia....: 0101                              @ Conta......: 74414-0\
          \                           @ Nome.......: EUGENIO SCHMITT COELHO      \
          \      @@------------------------------------------------@ Favorecido: EUGENIO\
          \ SCHMITT COELHO             @ Agencia Destino:                        \
          \   0101@ Conta Destino:         Conta Corrente - 74414-0@@Tipo        \
          \    NSU                        Valor@Dinheiro        697841472523     \
          \        R$20,00@------------------------------------------------@     \
          \          SAC - 0800 123 4569@      SAC - DEF. AUDITIVO - 0800 123 4568@\
          \            OUVIDORIA - 0800 123 4567@@             www.siteparceiro.com.br@"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
  emprestimoReq:
    type: "object"
    required:
    - "Cartao"
    - "InfEmprestimo"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cartao:
        $ref: "#/definitions/cartao"
      Cripto:
        $ref: "#/definitions/cripto"
      InfEmprestimo:
        $ref: "#/definitions/infEmprestimoReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  emprestimoResp:
    type: "object"
    required:
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      Cripto:
        hash: "hash"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
...
  identificacaoReq:
    type: "object"
    required:
    - "Cartao"
    - "Cripto"
    - "InfIdentificacao"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cartao:
        $ref: "#/definitions/cartao"
      Cripto:
        $ref: "#/definitions/cripto"
      InfIdentificacao:
        $ref: "#/definitions/infIdentificacaoReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
...
  terminal:
    type: "object"
    required:
    - "codEstab"
    - "id"
    - "tipo"
    properties:
      codEstab:
        type: "string"
        example: "000000000742673"
        description: "Código do estabelecimento onde está localizado o terminal gerado\
          \ pela Saque e Pague (15 dígitos)."
        pattern: "^\\d{15}$"
      id:
        type: "string"
        example: "05100004"
        description: "Identificador do terminal gerado pela Saque e Pague (8 dígitos)."
        pattern: "^\\d{8}$"
      tipo:
        type: "string"
        example: "008"
        description: "Tipo de terminal o qual a operação é gerada (3 dígitos, 008\
          \ = ATM)."
        pattern: "^\\d{3}$"
    description: "Informações sobre o terminal vigente."
    example:
      codEstab: "000000000742673"
      tipo: "008"
      id: "05100004"
  transacConf:
    type: "object"
    required:
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoConf"
      Terminal:
        $ref: "#/definitions/terminal"
...